Opprett oppgave + flytt funksjon til egen fil
